requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.0.0
Flask-Migrate==4.0.5
Flask-Talisman==1.1.0
Flask-Caching==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
itsdangerous==2.1.2
Jinja2==3.1.2
sentry-sdk==1.35.0
boto3==1.34.11
redis==5.0.1
Pillow==10.2.0
email-validator==2.1.0.post1
gunicorn==21.2.0
razorpay==1.4.1
